Removed redundant configuration contracts for business logic Added shared contracts for Inventory and Warehouse entities, to follow DRY principle Added more specific methods for warehouse entities Fixed some stylings in error interface
