+++ app/code/Magento/CatalogInventory/Api/Data/WarehouseConfigurationInterface.php
@@ -10,37 +10,22 @@ use Magento\Framework\Api\ExtensibleDataInterface;
  *
  * @api
  */
-interface WarehouseConfigurationInterface extends ExtensibleDataInterface
+interface WarehouseConfigurationInterface extends WarehouseRecordConfigurationDataInterface, ExtensibleDataInterface
 {
     /**
      * Sets inventory identifier for warehouse configuration
      *
-     * @param int $identifier
+     * @param int $warehouseId
      * @return $this
      */
-    public function setId($identifier);
+    public function setWarehouseId($warehouseId);
 
     /**
      * Returns related inventory identifier for configuration
      *
      * @return int
      */
-    public function getId();
-
-    /**
-     * Sets is managed flag
-     *
-     * @param int|null $flag
-     * @return $this
-     */
-    public function setIsManaged($flag);
-
-    /**
-     * Returns is managed flag
-     *
-     * @return int|null
-     */
-    public function getIsManaged();
+    public function getWarehouseId();
 
     /**
      * Sets allowed product type ids,
@@ -73,36 +58,6 @@ interface WarehouseConfigurationInterface extends ExtensibleDataInterface
      * @return int|null
      */
     public function getIsAllowedQuantityChange();
-
-    /**
-     * Sets flag for decimal values in quantity
-     *
-     * @param int|null $flag
-     * @return $this
-     */
-    public function setIsDecimalQuantity($flag);
-
-    /**
-     * Returns flag for decimal values in quantity
-     *
-     * @return int|null
-     */
-    public function getIsDecimalQuantity();
-
-    /**
-     * Sets flag for allowed reservation below minimum available quantity
-     *
-     * @param int|null $flag
-     * @return $this
-     */
-    public function setIsAllowedReservationBelowMinimumQty($flag);
-
-    /**
-     * Returns flag for allowed reservation below minimum available quantity
-     *
-     * @return int|null
-     */
-    public function getIsAllowedReservationBelowMinimumQty();
 
     /**
      * Sets flag for allowing return quantities for cancelled reservation
@@ -148,97 +103,6 @@ interface WarehouseConfigurationInterface extends ExtensibleDataInterface
      * @return int|null
      */
     public function getIsOutOfStockVisible();
-
-    /**
-     * Sets quantity for notification about low quantity level
-     *
-     * @param float|null $quantity
-     * @return $this
-     */
-    public function setLowStockNotificationQuantity($quantity);
-
-    /**
-     * Returns quantity for notification about low quantity level
-     *
-     * @return float|null
-     */
-    public function getLowStockNotificationQuantity();
-
-
-    /**
-     * Sets flag for enabled incremental quantity change
-     *
-     * @param int|null $flag
-     * @return $this
-     */
-    public function setIsAllowedQuantityIncrement($flag);
-
-    /**
-     * Returns flag for enabled incremental quantity change
-     *
-     * @return int|null
-     */
-    public function getIsAllowedQuantityIncrement();
-
-    /**
-     * Sets quantity increment step
-     *
-     * @param float|null
-     */
-    public function setQuantityIncrement($increment);
-
-    /**
-     * Returns quantity increment step
-     *
-     * @return float|null
-     */
-    public function getQuantityIncrement();
-
-    /**
-     * Minimum available quantity for warehouse
-     *
-     * @param float|null $quantity
-     * @return $this
-     */
-    public function setMinimumAvailableQuantity($quantity);
-
-    /**
-     * Minimum available quantity for warehouse
-     *
-     * @return float|null
-     */
-    public function getMinimumAvailableQuantity();
-
-    /**
-     * Sets minimum quantity for a reservation request in warehouse
-     *
-     * @param float|null $quantity
-     * @return $this
-     */
-    public function setMinimumReservationQuantity($quantity);
-
-    /**
-     * Returns minimum quantity for a reservation request in warehouse
-     *
-     * @return float|null
-     */
-    public function getMinimumReservationQuantity();
-
-
-    /**
-     * Sets maximum quantity for a reservation request in warehouse
-     *
-     * @param float|null $quantity
-     * @return $this
-     */
-    public function setMaximumReservationQuantity($quantity);
-
-    /**
-     * Returns maximum quantity for a reservation request in this inventory
-     *
-     * @return float|null
-     */
-    public function getMaximumReservationQuantity();
 
     /**
      * Sets availability status codes for warehouse configuration
